+++ src/components/options/Choices.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {Button, Dropdown, Form, Divider} from 'semantic-ui-react';
 import {connect} from 'react-redux';
 import {isGlobalValueSet} from '../../helpers/utils';
-import {createValueset, updateItem} from '../../actions';
+import {createValueset, updateItem, makeValuesetGlobal, copyValuesetLocal} from '../../actions';
 import ValueSetEditor from '../ValueSetEditor';
 
 class Choices extends Component {
@@ -13,17 +13,31 @@ class Choices extends Component {
           : [];
       return (
         <React.Fragment>
-          <Form.Field>
-            <label>Select global list</label>
-            <Dropdown fluid search selection options={options} value={this.props.item.get('valueSetId')}
-              onChange={(evt, data) => this.props.updateItem(this.props.item.get('id'), 'valueSetId', data.value)} />
-          </Form.Field>
+          
+          <div style={{display: 'flex', flexDirection: 'row', width: '100%', alignItems: 'flex-end'}}>
+            <div style={{paddingRight: '10px', flexGrow: 1}}>
+              <Form.Field>
+                <label>Select global list</label>
+                <Dropdown fluid search selection options={options} value={this.props.item.get('valueSetId')}
+                  onChange={(evt, data) => this.props.updateItem(this.props.item.get('id'), 'valueSetId', data.value)} />
+              </Form.Field>
+            </div>
+            <div>
+              <Button disabled={!this.props.item.get('valueSetId')} onClick={() => this.props.copyValuesetLocal(this.props.item.get('valueSetId'), this.props.item.get('id'))}>Copy as local</Button>
+            </div>
+          </div>
+         
           <Divider horizontal>Or</Divider>
           <Button onClick={() => this.props.createValueset(this.props.item.get('id'))}>Create local list</Button>
         </React.Fragment>
       );
     } else  {
-      return (<ValueSetEditor valueSetId={this.props.item.get('valueSetId')} />);
+      return (
+        <React.Fragment>
+          <ValueSetEditor valueSetId={this.props.item.get('valueSetId')} />
+          <Button onClick={() => this.props.makeValuesetGlobal(this.props.item.get('valueSetId'))}>Make Global</Button>
+        </React.Fragment>
+      );
     }
   }
 }
@@ -33,7 +47,9 @@ const ChoicesConnected = connect(
     globalValueSets: state.dialobComposer.form && state.dialobComposer.form.getIn(['metadata', 'composer', 'globalValueSets'])
   }), {
     createValueset,
-    updateItem
+    updateItem,
+    makeValuesetGlobal,
+    copyValuesetLocal
   }
 )(Choices);